+++ src/main/java/it/pagopa/pn/delivery/springbootcfg/MDCTraceIdActivation.java
@@ -1,0 +1,10 @@
+package it.pagopa.pn.delivery.springbootcfg;
+
+
+import it.pagopa.pn.commons.log.MDCWebFilter;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MDCTraceIdActivation extends MDCWebFilter {
+
+}
